+++ cli/lsp/documents/to_lsp_range.rs
@@ -1,0 +1,14 @@
+use tower_lsp::lsp_types as lsp;
+
+pub fn to_lsp_range(range: &deno_graph::Range) -> lsp::Range {
+  lsp::Range {
+    start: lsp::Position {
+      line: range.start.line as u32,
+      character: range.start.character as u32,
+    },
+    end: lsp::Position {
+      line: range.end.line as u32,
+      character: range.end.character as u32,
+    },
+  }
+}
